Naming Conventions: fix margin bug, remove last use of FontAwesome icon
This is the only page that still uses a FontAwesome icon. With this
removed, we can freely change how icons in the skin itself are
shown within the jquery-wp-content repo, without worrying about content
using it as well.

Fix a bug where `row` is applied too broadly, causing the heading to
go into negative margin.